Add files via upload
All the info is hardcoded right now. Still have to find a way to make the data change based on spreadsheet. Also need to do a verification (or add the data if the email isn't on the spreadsheet) and after that, will go to profile page. Currently the profile page is not accessible by the nav bar.